Move scatter_dense_block; tidy imports & ipc fix
Move the scatter_dense_block utility from fedoo/time/common.py to fedoo/core/_sparsematrix.py and update consumers to import it from the core module (generalized_alpha.py, rigid_body.py). Remove the duplicate implementation and the scipy.sparse import from time/common. Also: import ipctk directly in ipc_contact.py to avoid re-running the version check on the contact hot path, update a doc comment in rigid_body.py to reference RigidTie.rotation_jacobian(), and simplify the static_regularisation initialization by removing the mass scaling calculation.
